src/shared/requirements.txt
flask==2.3.3
requests==2.31.0
pillow==10.0.1
werkzeug==2.3.7 